requirements.txt
torch==1.9.0
opencv-python==4.5.3.56
pyyaml==5.4.1
numpy==1.21.2
scikit-learn==0.24.2